feat(admin-app): package the admin client runtime glue (#3044 Phase 2)
Phase 2 slice 1 of the source-free admin host. Add
`@voyant-travel/admin-app/runtime`, the profile-agnostic client glue the admin
routes/router need, lifted out of the operator starter:

- `managedProfileAdminFetcher` — the isomorphic Voyant fetcher (cookie-forward
  on the server, admin-path normalize on the client). `createIsomorphicFn` stays
  at module scope so the TanStack Start build strips the `.server(...)` branch
  from client bundles (verified: no `getRequestHeader`/`react-start/server` in
  dist/client).
- `normalizeAdminApiUrl(url)` — `/v1/<module>` → `/v1/admin/<module>` rewrite.
- `getManagedProfileAdminApiUrl()` — same-origin `/api` base URL (reads
  `process` off `globalThis` so the browser-targeted package needs no node types).

The starter's `lib/voyant-fetcher.ts` / `env.ts` / `operator-admin-api-paths.ts`
become thin re-export shims, preserving their export names so their ~31
importers are unaffected. Also records the Phase 2 investigation findings in the
plan doc (glue is client-side; the `voyant admin generate` codegen lives in the
separate cli repo, making the route-map slice cross-repo; auth is a thick
starter stack).

Verified: admin-app + operator typecheck, admin-app tests (22), operator builds
and boots, `/` renders the SSR shell (200) exercising the server fetcher branch,
`/api/health` 200.
